Read and save a mark from the assessment sheet

+++ grading/models.py
@@ -61,7 +61,8 @@ class Student(models.Model):
 class Report(models.Model):
     assignment = models.ForeignKey(Assignment, on_delete=models.CASCADE)
     student = models.ForeignKey(Student, on_delete=models.CASCADE)
-    mark = models.IntegerField(blank=True, null=True)
+    mark = models.DecimalField(max_digits=2, decimal_places=1, blank=True,
+                               null=True)
     report = models.FileField()
     assessment = models.FileField()
